Classify observations.

club() is used to classify obervations using binary procrustes rotation.

Usage

club(
  f,
  data,
  imprecision,
  nreps,
  normalise_cols,
  reorder_obs,
  display_progress
)

Arguments

f

a formula.

data

a data.frame.

imprecision

a number indicting the margin of imprecision allowed in classification.

nreps

the number of replicates to use in the randomisation test.

normalise_cols

a boolean indicating whether to normalise matrix columns.

reorder_obs

a string indicating the method for reordering observations to calculate c-values.

display_progress

a boolean indictaing whether a progress bar should be displayed.

Value

an object of class "clubprofit" is a list containing the folllowing components:

prediction

a character vector of predicted classifications.

accuracy

a character vector indicating whether each classification is "correct", "incorrect", or "ambiguous".

pcc

the percentage of correct classifications.

cval

the chance of randomly reordered data producing a PCC >= the observed PCC.

pcc_replicates

a vector of PCCs generated from randomly reordered data used to calculate cval.

call

the matched call.

Examples

mod <- club(rate ~ dose, data = caffeine)