Individual level classification results.

Individual level classification results.

Usage

individual_results(m, digits)

Arguments

m

an object of class "clubprofit" produced by club()

digits

an integer

Value

a data.frame containing a columns of predictions and prediction accuracy

Details

Returns a data.frame containing predicted classifications and classification accuracy for each individual observation.

Examples

mod <- club(rate ~ dose, data = caffeine)
individual_results(mod)
#>    individual observation target prediction  accuracy  csi
#> 1           1         242      0          0   correct 1.00
#> 2           2         245      0        100 incorrect 0.64
#> 3           3         244      0          0   correct 0.87
#> 4           4         248      0        200 incorrect 0.67
#> 5           5         247      0        100 incorrect 0.91
#> 6           6         248      0        200 incorrect 0.67
#> 7           7         242      0          0   correct 1.00
#> 8           8         244      0          0   correct 0.87
#> 9           9         246      0        100 incorrect 0.73
#> 10         10         242      0          0   correct 1.00
#> 11         11         248    100        200 incorrect 0.67
#> 12         12         246    100        100   correct 0.73
#> 13         13         245    100        100   correct 0.64
#> 14         14         247    100        100   correct 0.91
#> 15         15         248    100        200 incorrect 0.67
#> 16         16         250    100        200 incorrect 0.93
#> 17         17         247    100        100   correct 0.91
#> 18         18         246    100        100   correct 0.73
#> 19         19         243    100        100   correct 1.00
#> 20         20         244    100          0 incorrect 0.87
#> 21         21         246    200        100 incorrect 0.73
#> 22         22         248    200        200   correct 0.67
#> 23         23         250    200        200   correct 0.93
#> 24         24         252    200        200   correct 1.00
#> 25         25         248    200        200   correct 0.67
#> 26         26         250    200        200   correct 0.93
#> 27         27         246    200        100 incorrect 0.73
#> 28         28         248    200        200   correct 0.67
#> 29         29         245    200        100 incorrect 0.64
#> 30         30         250    200        200   correct 0.93