Generate a summary of results from a fitted clubpro model.

Generate a summary of results from a fitted clubpro model.

Usage

# S3 method for class 'clubprofit'
summary(object, ...)

Arguments

object

an object of class "clubprofit".

...

ignored

Value

No return value, called for side effects.

Examples

mod <- club(rate ~ dose, data = caffeine)
summary(mod)
#> ********** Model Summary **********
#> 
#> ----- Classification Results -----
#> Observations:  30 
#> Missing observations:  0 
#> Target groups:  3 
#> Correctly classified observations:  18 
#> Incorrectly classified observations:  12 
#> Ambiguously classified observations:  0 
#> PCC:  60 
#> Median classification strength index:  0.8 
#> 
#> ----- Randomisation Test Results -----
#> Random reorderings:  1000 
#> Minimum random PCC:  30 
#> Maximum random PCC:  76.67 
#> Chance-value:  0.53